requirements.txt
pytz
django-mptt
djangorestframework
django-ckeditor
django-codemirror2
django-mptt-graph
jsonfield
django-jsonedutor